client/requirements.txt
grpcio
grpcio-tools
protobuf
PyYAML
python-dotenv
reedsolo
tenacity
tqdm
colorlog